requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.7
django-autoslug==1.9.9
django-bootstrap-modal-forms==2.2.0
django-filter==22.1
django-js-asset==2.0.0
django-mptt==0.13.4
django-mptt-admin==2.3.0
django-property-filter==1.1.1
django-singleton==0.1.9
django-widget-tweaks==1.4.12
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
docutils==0.19
drf-yasg==1.21.5
exceptiongroup==1.1.0
generic==1.1.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
python-dotenv==1.0.0
requests==2.28.1
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.14

